Fix for uart flush, power down button handling changes, interrupt functions for gpio driver
Added fix to drv_uart_flush to fix bug with inability to talk to
quintics when the brain first boots.
Changed power button handling so that when user presses and holds the
button, it powers down before the button is released.
Added configure interrupt functions that are used the change the
functionality of the gpio interrupts.