Added commands to control brain pack remotely. Improved reconnection and missing NOD detection.
Added commands that can be sent to the brain pack that emulate the
buttons on it.
Also added missing packet tracking for specific IMUs to detect and
rectify a disconnection more easily.
Added separate debugPrint function that is disabled by default so that
the debug prints don't mess up the data stream.
Added some changes made by Hriday to support the fabric sense