Changed and integrated the Debug Log task as a SD-card task. Added fix for the sensors.
The system can record the Debug Log prints to DebugLog.txt file. There
is a check on the size of the file and it is backed up if it exceeds the
size.
For the sensors, added in the fix which puts the system to error if the
function fails to initialize.