Added the Debug Log task as a part of SD-card task. Added in fixes for the sensors.
The system can record the Debug Log prints to DebugLog.txt file. There
is a check on the size of the file and it is backed up if it exceeds the
size.
For the sensors, added in the fix which puts the system to error if the
function fails to initialize.