Fix for SD card buffer write full issue
Added fix for the brainpack failing to write to the buffer in certain
instances. This was caused by having the buffer too small which did not
take into account the variability of the SD card write times.
Re-added a power cycling of the NODs before reset as it has been shown
that the NODs go into a fast advertising state after power is
disconnected. This should improve the connection time.