Added functionality to store and load serial number of brain pack from NVM (flash) memory, driving of LBO from STAT pin and cycling of JcEn
The Serial number is stored in NVM by a command once in a brainpack
lifetime and is loaded to to RAM every time it turns on. The LBO is now
driven by STAT pin and flashes red LED thrice when battery is below a
limit,  the brain pack doesn't turn on until charged. The Jack Enables
are cycled until a jack is detected.